In progress of adding address suggestions and added error handling
